added modules for actions, moved to use new AuthServiceProviderRegistry, removed API Flags, general cleanup
