+++ portability-gateway/src/main/java/org/dataportabilityproject/gateway/action/listdatatypes/ListDataTypesAction.java
@@ -18,7 +18,7 @@ package org.dataportabilityproject.gateway.action.listdatatypes;
 import com.google.inject.Inject;
 import java.util.Set;
 import org.dataportabilityproject.gateway.action.Action;
-import org.dataportabilityproject.spi.transfer.provider.TransferServiceProviderRegistry;
+import org.dataportabilityproject.spi.gateway.auth.AuthServiceProviderRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,11 +32,11 @@ public final class ListDataTypesAction implements
   private static final Logger logger = LoggerFactory.getLogger(
       ListDataTypesAction.class);
 
-  private final TransferServiceProviderRegistry registry;
+  private final AuthServiceProviderRegistry registry;
 
   @Inject
   ListDataTypesAction(
-      TransferServiceProviderRegistry registry
+      AuthServiceProviderRegistry registry
   ) {
     this.registry = registry;
   }
