+++ portability-spi-transfer/src/main/java/org/dataportabilityproject/spi/transfer/provider/TransferServiceProviderRegistry.java
@@ -1,6 +1,5 @@
 package org.dataportabilityproject.spi.transfer.provider;
 
-import java.util.Set;
 import org.dataportabilityproject.types.transfer.PortableType;
 import org.dataportabilityproject.types.transfer.auth.AuthData;
 import org.dataportabilityproject.types.transfer.models.DataModel;
@@ -19,6 +18,7 @@ public interface TransferServiceProviderRegistry {
    */
   Exporter<?, ?> getExporter(String serviceId, String transferDataType);
 
+
   /**
    * Returns the exporter that supports the serviceId and transferDataType.
    *
@@ -26,16 +26,4 @@ public interface TransferServiceProviderRegistry {
    * @param transferDataType the transfer data type
    */
   Importer<?, ?> getImporter(String serviceId, String transferDataType);
-
-  /**
-   * Returns the set of service ids that can transfered for the given {@code transferDataType}.
-   *
-   * @param transferDataType the transfer data type
-   */
-  Set<String> getServices(String transferDataType);
-
-  /**
-   * Returns the set of data types that support both import and export.
-   */
-  Set<String> getTransferDataTypes();
 }
